Update README.md
The particle size of metal-based catalysts and the accurate prediction of Platinum on Carbon (Pt/Carbon) electrocatalysts
are crucial for optimizing the efficiency of electrocatalysts used in the hydrogen evolution reaction. Existing methods often struggle with the complex, non-linear relationships between particle size and catalytic performance, and predicting the
performance of Pt/carbon electrocatalysts remains limited in terms of both the performance and reliability. In this study, we
propose categorical boosting (CatBoost) and extreme gradient boosting (XGBoost) models to enhance the optimization of
metal-based catalyst particle sizes and to precisely predict the performance of Pt/carbon electrocatalysts, respectively. Our
analysis reveals a linear relationship between the measured overpotential and the particle size of metal-based catalysts, with
the optimal particle size ranging from 1 to 5.0 nm. The CatBoost model outperforms the baseline gradient boosting regressor
(GBR) model achieved a 2.98% increase in R
2
and reductions of 24.67% in mean absolute error (MAE) and 47.56% in mean
squared error (MSE). The XGBoost model showed a 1.14% improvement in R
2
, and reduced MAE by 21.11% and MSE by
20.80%, demonstrating strong predictive performance in acidic electrolytes compared to the GBR model. In addition, SHapley Additive exPlanations (SHAP) and moDel Agnostic Language for Exploration and eXplanation (Dalex) analyses provide
deeper insights into the factors influencing catalytic performance.